Added Delete Videos Button
