+++ apps/desktop/src/utilities.js
